requirements.txt
discord==2.3.2
discord.py==2.7.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-auth-oauthlib==1.3.0
greenlet==3.3.2
gspread==6.2.1
idna==3.11
toml==0.10.2
